+++ Integrations/JiraV2/CHANGELOG.md
@@ -1,6 +1,13 @@
 ## [Unreleased]
 
 
+## [20.2.0] - 2020-02-04
+Fixed an issue in ***jira-get-issue*** where trying to get the attachment of the issue would fail.
+
+## [20.1.2] - 2020-01-22
+Fixed the description of the *reporter* argument in the ***jira-create-issue*** command.
+Fixed an issue where trying to fetch incidents when the idOffset was not configured would raise an error.
+
 ## [19.12.0] - 2019-12-10
 - API token parameter is now encrypted.
 - Improved authentication methods support - Basic, OAuth 1.0 (only from 5.0)