+++ Scripts/DBotPredictPhishingWords/CHANGELOG.md
@@ -1,6 +1,15 @@
 ## [Unreleased]
 
 
+## [20.2.3] - 2020-02-18
+-
+
+## [20.1.2] - 2020-01-22
+Added the argument *emailBodyHTML*, which enables you to pass the raw HTML of the email body.
+
+## [20.1.0] - 2020-01-07
+    -
+
 ## [19.12.1] - 2019-12-25
     - Added Support for highlighting in the text.
     - Support min text length argument